Update patch set 3
Patch Set 3:

(1 comment)

The issue was that the releases job was not able to find the autorelease package wanting to be released. 
This release is wrong too but unfortunately the change is merged. 
The job mentioned in the file produced 1.4.0 artifacts, not 1.4.1
https://jenkins.onap.org/job/sdc-sdc-distribution-client-maven-stage-master/85/console

Patch-set: 3
Reviewer: Gerrit User 481 <481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0