Enable case insensitive matching of http headers in distribution client
This change creates simple case insensitive map implementation that
is used in HttpSdcResponse class to store the response headers.
Generally HTTP headers are supposed to be case insensitive and
sometimes get transformed (in our case - by istio sidecar) to lower case.
With this change lower case will still be matched and the original
behaviour will not break.

Issue-ID: SDC-4233
Change-Id: I1dbaa6a65b9e993b27edfcbf569236328740d0e4
Signed-off-by: Marek Szwałkiewicz <marek.szwalkiewicz@external.t-mobile.pl>